requirements.txt
dnspython==2.3.0
email-validator==1.3.1
greenlet==2.0.2
idna==3.4
MarkupSafe==2.1.2
pydantic==1.10.5
PyJWT==2.6.0
PyMySQL==1.0.2
python-decouple==3.7
python-dotenv==0.21.1
SQLAlchemy==2.0.4
typing-extensions==4.5.0
Werkzeug==2.2.3